Move email-regex to common package
