+++ app/src/main/java/com/hedvig/app/feature/embark/masking/Unmasking.kt
@@ -1,6 +1,6 @@
 package com.hedvig.app.feature.embark.masking
 
-import com.hedvig.app.feature.embark.BIRTH_DATE_REVERSE
+import com.hedvig.app.feature.embark.util.BIRTH_DATE_REVERSE
 import java.time.LocalDate
 
 fun unmask(text: String, mask: String?): String = when (mask) {
