+++ app/src/main/java/com/hedvig/app/util/Regex.kt
@@ -1,0 +1,5 @@
+package com.hedvig.app.util
+
+import java.util.regex.Pattern
+
+val EMAIL_REGEX: Pattern = Pattern.compile("^.+@.+\\..+\$", Pattern.CASE_INSENSITIVE)
